Create Main2
There are four quadrants, numbered from 1 to 4, as shown in the diagram below:image

For example, the point A, which is at coordinates (12,5) lies in quadrant 1 since both its x and y values are positive, and point B lies in quadrant 2 since its x value is negative and its y value is positive.

Your job is to take a point and determine the quadrant it is in. You can assume that neither of the two coordinates will be 0.

Input Format

The first line of input contains the integer x.

The second line of input contains the integer y.